+++ declavatar/src/decl_v2/sexpr/da.rs
@@ -2,7 +2,8 @@ pub(super) mod asset;
 pub(super) mod avatar;
 pub(super) mod controller;
 pub(super) mod driver;
-pub(super) mod layer;
+pub(super) mod layer_basic;
+pub(super) mod layer_raw;
 pub(super) mod menu;
 pub(super) mod parameter;
 
@@ -15,7 +16,8 @@ pub fn define_da_module(scope: Scope) -> Module {
     parameter::register_parameter_function(&scope);
     asset::register_asset_function(&scope);
     controller::register_controller_function(&scope);
-    layer::register_layer_function(&scope);
+    layer_basic::register_layer_basic_function(&scope);
+    layer_raw::register_layer_raw_function(&scope);
     driver::register_driver_function(&scope);
     menu::register_menu_function(&scope);
