+++ declavatar/src/decl_v2/sexpr/da/layer_basic.rs
@@ -4,10 +4,7 @@ use crate::decl_v2::{
         layer::{
             DeclControllerLayer, DeclGroupCopyMode, DeclGroupLayer, DeclGroupMaterialTarget,
             DeclGroupObjectTarget, DeclGroupOption, DeclGroupOptionKind, DeclGroupOptionTarget,
-            DeclGroupShapeTarget, DeclLayerInlineAnimation, DeclPuppetLayer, DeclRawLayer,
-            DeclRawLayerAnimation, DeclRawLayerAnimationKind, DeclRawLayerBlendTreeField,
-            DeclRawLayerBlendTreeType, DeclRawLayerState, DeclRawLayerTransition,
-            DeclRawLayerTransitionCondition, DeclRawLayerTransitionOrdering, DeclSwitchLayer,
+            DeclGroupShapeTarget, DeclPuppetLayer, DeclSwitchLayer,
         },
         StaticTypeName,
     },
@@ -20,7 +17,7 @@ use crate::decl_v2::{
 
 use ketos::{Arity, Error, ExecError, Name, NameStore, Scope, Value};
 
-pub fn register_layer_function(scope: &Scope) {
+pub fn register_layer_basic_function(scope: &Scope) {
     // layer functions
     register_function(
         scope,
@@ -43,13 +40,6 @@ pub fn register_layer_function(scope: &Scope) {
         Arity::Min(1),
         &["driven-by", "default-mesh", "animation"],
     );
-    register_function(
-        scope,
-        "raw-layer",
-        declare_raw_layer,
-        Arity::Min(1),
-        &["default"],
-    );
 
     // option functions
     register_function(
@@ -59,14 +49,6 @@ pub fn register_layer_function(scope: &Scope) {
         Arity::Min(1),
         &["value", "animation"],
     );
-    register_function(
-        scope,
-        "inline-animation",
-        declare_inline_animation,
-        Arity::Min(0),
-        &[],
-    );
-    register_function(scope, "state", declare_state, Arity::Min(2), &[]);
 
     // set-x functions
     register_function(
@@ -89,36 +71,6 @@ pub fn register_layer_function(scope: &Scope) {
         declare_set_material,
         Arity::Exact(2),
         &["mesh"],
-    );
-
-    // raw layer functions
-    register_function(
-        scope,
-        "clip",
-        declare_clip,
-        Arity::Min(1),
-        &["speed", "speed-by", "time-by"],
-    );
-    register_function(
-        scope,
-        "blendtree",
-        declare_blendtree,
-        Arity::Min(0),
-        &["type", "x", "y"],
-    );
-    register_function(
-        scope,
-        "blendtree-field",
-        declare_blendtree_field,
-        Arity::Range(2, 3),
-        &[],
-    );
-    register_function(
-        scope,
-        "transition-to",
-        declare_transition_to,
-        Arity::Min(1),
-        &["duration"],
     );
 }
 
@@ -250,30 +202,6 @@ fn declare_puppet_layer(
         default_mesh: default_mesh.map(|dm| dm.to_string()),
         animation_asset: animation_asset.map(|a| a.to_string()),
         keyframes,
-    })
-    .into())
-}
-
-fn declare_raw_layer(
-    _name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let name: &str = args.exact_arg(function_name, 0)?;
-    let default: Option<&str> = args.exact_kwarg("default")?;
-
-    let mut states = vec![];
-    for state_value in args.args_after_recursive(function_name, 1)? {
-        states.push(
-            state_value
-                .downcast_foreign_ref::<&DeclRawLayerState>()?
-                .clone(),
-        );
-    }
-    Ok(DeclControllerLayer::Raw(DeclRawLayer {
-        name: name.to_string(),
-        default: default.map(|d| d.to_string()),
-        states,
     })
     .into())
 }
@@ -425,249 +353,4 @@ fn declare_set_material(
         mesh: mesh.map(|m| m.to_string()),
     }
     .into())
-}
-
-fn declare_state(
-    _name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let name: &str = args.exact_arg(function_name, 0)?;
-    let kind: &DeclRawLayerAnimationKind = args.exact_arg(function_name, 1)?;
-
-    let mut transitions = vec![];
-    for transition_value in args.args_after_recursive(function_name, 2)? {
-        transitions.push(
-            transition_value
-                .downcast_foreign_ref::<&DeclRawLayerTransition>()?
-                .clone(),
-        );
-    }
-
-    Ok(DeclRawLayerState {
-        name: name.to_string(),
-        kind: kind.clone(),
-        transitions,
-    }
-    .into())
-}
-
-fn declare_inline_animation(
-    _name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let mut targets = vec![];
-    for target_value in args.args_after_recursive(function_name, 0)? {
-        targets.push(take_option_target(target_value)?);
-    }
-
-    Ok(DeclLayerInlineAnimation { targets }.into())
-}
-
-fn declare_clip(
-    _name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let animation: &Value = args.exact_arg(function_name, 0)?;
-    let speed: Option<f64> = args.exact_kwarg("speed")?;
-    let speed_by: Option<&str> = args.exact_kwarg("speed-by")?;
-    let time_by: Option<&str> = args.exact_kwarg("time-by")?;
-
-    Ok(DeclRawLayerAnimationKind::Clip {
-        animation: take_animation(animation)?,
-        speed: (speed, speed_by.map(|s| s.to_string())),
-        time: time_by.map(|t| t.to_string()),
-    }
-    .into())
-}
-
-fn declare_blendtree(
-    name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let blend_type: &str = match args.exact_kwarg_expect::<&Value>("type")? {
-        Value::Name(s) => name_store.get(*s),
-        v => {
-            return Err(Error::ExecError(ExecError::TypeError {
-                expected: "blendtree type name",
-                found: v.type_name(),
-                value: Some(v.clone()),
-            }))
-        }
-    };
-    let tree_type = match blend_type {
-        "linear" => {
-            let x: &str = args.exact_kwarg_expect("x")?;
-            DeclRawLayerBlendTreeType::Linear(x.to_string())
-        }
-        "simple-2d" => {
-            let x: &str = args.exact_kwarg_expect("x")?;
-            let y: &str = args.exact_kwarg_expect("y")?;
-            DeclRawLayerBlendTreeType::Simple2D(x.to_string(), y.to_string())
-        }
-        "freeform-2d" => {
-            let x: &str = args.exact_kwarg_expect("x")?;
-            let y: &str = args.exact_kwarg_expect("y")?;
-            DeclRawLayerBlendTreeType::Freeform2D(x.to_string(), y.to_string())
-        }
-        "cartesian-2d" => {
-            let x: &str = args.exact_kwarg_expect("x")?;
-            let y: &str = args.exact_kwarg_expect("y")?;
-            DeclRawLayerBlendTreeType::Cartesian2D(x.to_string(), y.to_string())
-        }
-        _ => {
-            return Err(Error::Custom(
-                DeclSexprError::KeywordExpected("blendtree type name".to_string()).into(),
-            ))
-        }
-    };
-
-    let mut fields = vec![];
-    for field_value in args.args_after_recursive(function_name, 0)? {
-        fields.push(
-            field_value
-                .downcast_foreign_ref::<&DeclRawLayerBlendTreeField>()?
-                .clone(),
-        );
-    }
-
-    Ok(DeclRawLayerAnimationKind::BlendTree { tree_type, fields }.into())
-}
-
-fn declare_blendtree_field(
-    _name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let animation: &Value = args.exact_arg(function_name, 0)?;
-    let x_value: Option<f64> = args.try_exact_arg(1)?;
-    let y_value: Option<f64> = args.try_exact_arg(2)?;
-
-    Ok(DeclRawLayerBlendTreeField {
-        animation: take_animation(animation)?,
-        values: [x_value.unwrap_or(0.0), y_value.unwrap_or(0.0)],
-    }
-    .into())
-}
-
-fn take_animation(animation_value: &Value) -> KetosResult<DeclRawLayerAnimation> {
-    let target = match animation_value.type_name() {
-        "string" => {
-            let Value::String(s) = animation_value else {
-                unreachable!("must be string")
-            };
-            DeclRawLayerAnimation::External(s.to_string())
-        }
-        DeclLayerInlineAnimation::TYPE_NAME => DeclRawLayerAnimation::Inline(
-            animation_value
-                .downcast_foreign_ref::<&DeclLayerInlineAnimation>()?
-                .clone(),
-        ),
-        _ => {
-            return Err(Error::Custom(
-                DeclSexprError::UnexpectedTypeValue(
-                    animation_value.type_name().to_string(),
-                    "string or inline animation".to_string(),
-                )
-                .into(),
-            ))
-        }
-    };
-
-    Ok(target)
-}
-
-fn declare_transition_to(
-    name_store: &NameStore,
-    function_name: Name,
-    args: SeparateArguments,
-) -> KetosResult<Value> {
-    let target: &str = args.exact_arg(function_name, 0)?;
-    let duration: Option<f64> = args.exact_kwarg("duration")?;
-
-    // conditon is list, cannot be recursive
-    let mut and_terms = vec![];
-    for condition_value in args.args_after(function_name, 1)? {
-        let Value::List(condition_list) = condition_value else {
-            return Err(Error::ExecError(ExecError::TypeError {
-                expected: "list that expresses condition",
-                found: condition_value.type_name(),
-                value: Some((*condition_value).clone()),
-            }));
-        };
-        let condition = parse_condition(name_store, condition_list)?;
-        and_terms.push(condition);
-    }
-
-    Ok(DeclRawLayerTransition {
-        target: target.to_string(),
-        duration,
-        and_terms,
-    }
-    .into())
-}
-
-fn parse_condition(
-    name_store: &NameStore,
-    condition_list: &[Value],
-) -> KetosResult<DeclRawLayerTransitionCondition> {
-    if condition_list.len() != 3 {
-        return Err(Error::Custom(DeclSexprError::InvalidCondition.into()));
-    }
-
-    let parameter = match &condition_list[1] {
-        Value::String(p) => p.to_string(),
-        other => {
-            return Err(Error::ExecError(ExecError::TypeError {
-                expected: "string",
-                found: other.type_name(),
-                value: Some(other.clone()),
-            }))
-        }
-    };
-    let ordering = match &condition_list[0] {
-        Value::Name(n) => match name_store.get(*n) {
-            "=" => DeclRawLayerTransitionOrdering::Equal,
-            "/=" => DeclRawLayerTransitionOrdering::NotEqual,
-            ">" => DeclRawLayerTransitionOrdering::Greater,
-            "<" => DeclRawLayerTransitionOrdering::Lesser,
-            _ => return Err(Error::Custom(DeclSexprError::InvalidCondition.into())),
-        },
-        other => {
-            return Err(Error::ExecError(ExecError::TypeError {
-                expected: "operator",
-                found: other.type_name(),
-                value: Some(other.clone()),
-            }))
-        }
-    };
-    let condition = match &condition_list[2] {
-        Value::Bool(bv) => match ordering {
-            DeclRawLayerTransitionOrdering::Equal => {
-                DeclRawLayerTransitionCondition::Bool(parameter, *bv)
-            }
-            DeclRawLayerTransitionOrdering::NotEqual => {
-                DeclRawLayerTransitionCondition::Bool(parameter, !*bv)
-            }
-            _ => return Err(Error::Custom(DeclSexprError::InvalidCondition.into())),
-        },
-        Value::Integer(iv) => DeclRawLayerTransitionCondition::Int(
-            parameter,
-            ordering,
-            iv.to_i64()
-                .ok_or_else(|| Error::Custom(DeclSexprError::InvalidCondition.into()))?,
-        ),
-        Value::Float(fv) => DeclRawLayerTransitionCondition::Float(parameter, ordering, *fv),
-        other => {
-            return Err(Error::ExecError(ExecError::TypeError {
-                expected: "int, bool, or float",
-                found: other.type_name(),
-                value: Some(other.clone()),
-            }))
-        }
-    };
-    Ok(condition)
 }
